requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-migrate==4.0.4
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1
gunicorn==20.1.0
pytest==7.3.1
werkzeug==2.2.3
PyJWT==2.8.0
